+++ src/modules-arch/schemes/conto.js
@@ -25,6 +25,7 @@ import { MyAssetsExplorerModule } from '@/vue/modules/assets/my-assets-explorer/
 import { CustomersListPageModule } from '@/vue/pages/customers-list-page-module'
 import { BusinessesPageModule } from '@/vue/pages/businesses-page'
 import { BusinessesAllPageModule } from '@/vue/pages/businesses-all-page-module'
+import { BusinessesMyPageModule } from '@/vue/pages/businesses-my-page-module'
 import { CurrentBusinessIndicatorModule } from '@/vue/navigation/navbar/current-business-indicator/module'
 import { BusinessOwnershipModule } from '@/vue/navigation/navbar/business-ownership/module'
 import { AtomicSwapsPageModule } from '@/vue/pages/atomic-swaps-page-module'
@@ -92,6 +93,14 @@ export default {
             routerEntry: {
               path: '/businesses/all',
               name: vueRoutes.allBusinesses.name,
+              props: true,
+            },
+            isUnverifiedOnly: true,
+          }),
+          new BusinessesMyPageModule({
+            routerEntry: {
+              path: '/businesses/my',
+              name: vueRoutes.myBusinesses.name,
               props: true,
             },
             isUnverifiedOnly: true,