+++ src/vue/pages/businesses-page.js
@@ -1,5 +1,6 @@
 import { PageModuleDescriptor } from '@/modules-arch/page-module-descriptor'
 import { BusinessesAllPageModule } from '@/vue/pages/businesses-all-page-module'
+import { BusinessesMyPageModule } from '@/vue/pages/businesses-my-page-module'
 
 export class BusinessesPageModule extends PageModuleDescriptor {
   constructor (opts = {}) {
@@ -8,6 +9,7 @@ export class BusinessesPageModule extends PageModuleDescriptor {
       importComponentFn: _ => import('@/vue/pages/Businesses'),
       allowedSubmodules: [
         BusinessesAllPageModule,
+        BusinessesMyPageModule,
       ],
     })
   }